requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.4
Django==4.2.6
django-cors-headers==4.3.0
django-rest-knox==4.2.0
django-sendgrid-v5==1.2.2
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.2.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3.post1
requests==2.31.0
sendgrid==6.10.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
